+++ TransactionsService/Services/Implementations/TransactionService.cs
@@ -7,21 +7,22 @@ using TransactionsService.Models.Errors;
 using TransactionsService.Services.Mappings;
 using TransactionsService.Repositories.Implementations;
 using Microsoft.EntityFrameworkCore;
+using TransactionsService.Messaging;
+using TransactionsService.Messaging.Events;
+using TransactionsService.Models;
 
 namespace TransactionsService.Services.Implementations
 {
     public class TransactionService : ITransactionService
     {
         private readonly ITransactionRepository _transactionRepository;
-        private readonly IValidator<CreateTransactionRequest> _createTransactionRequestValidator;
-        private readonly IValidator<UpdateTransactionRequest> _updateTransactionRequestValidator;
+        private readonly IKafkaProducer _kafkaProducer;
+        
 
-        public TransactionService(ITransactionRepository transactionRepository, IValidator<CreateTransactionRequest> createTransactionRequestValidator,
-            IValidator<UpdateTransactionRequest> updateTransactionRequestValidator)
+        public TransactionService(ITransactionRepository transactionRepository, IKafkaProducer kafkaProducer)
         {
             _transactionRepository = transactionRepository;
-            _createTransactionRequestValidator = createTransactionRequestValidator;
-            _updateTransactionRequestValidator = updateTransactionRequestValidator;
+            _kafkaProducer = kafkaProducer;
         }
 
         public async Task<Result<TransactionResponse>> GetTransactionByIdAsync(long id, CancellationToken token)
@@ -36,14 +37,14 @@ namespace TransactionsService.Services.Implementations
         public async Task<Result<TransactionResponse>> CreateTransactionAsync(CreateTransactionRequest createTransactionRequest,
             CancellationToken token)
         {
-            var validationResult = await _createTransactionRequestValidator.ValidateAsync(createTransactionRequest, token);
-
-            if (!validationResult.IsValid)
-                return Result<TransactionResponse>.Failure(TransactionErrors.InvalidCredentials);
 
             var transaction = createTransactionRequest.ToTransaction();
 
             await _transactionRepository.AddAsync(transaction, token);
+
+            var transactionCreatedEvent = new TransactionEvent("transaction-created", new TransactionData(transaction));
+
+            await _kafkaProducer.ProduceAsync("transaction-events", transaction.Id.ToString(), transactionCreatedEvent);
 
             return Result<TransactionResponse>.Success(transaction.ToTransactionResponse());
         }
@@ -51,24 +52,40 @@ namespace TransactionsService.Services.Implementations
         public async Task<Result<TransactionResponse>> UpdateTransactionAsync(UpdateTransactionRequest updateTransactionRequest,
             CancellationToken token)
         {
-            var validationResult = await _updateTransactionRequestValidator.ValidateAsync(updateTransactionRequest, token);
-
-            if (!validationResult.IsValid)
-                return Result<TransactionResponse>.Failure(TransactionErrors.InvalidCredentials);
-
             var transaction = updateTransactionRequest.ToTransaction();
+            var prevValue = await _transactionRepository.GetValueByIdAsync(transaction.Id, token);
+            var prevCurrency = await _transactionRepository.GetCurrencyByIdAsync(transaction.Id, token);
             transaction = await _transactionRepository.UpdateAsync(transaction, token);
 
-            return transaction is null
-                ? Result<TransactionResponse>.Failure(TransactionErrors.TransactionNotFound)
-                : Result<TransactionResponse>.Success(transaction.ToTransactionResponse());
+            if(transaction != null && prevValue!=null && prevCurrency!=null)
+            {
+                var transactionUpdatedEvent = new TransactionEvent("transaction-updated", new TransactionData(transaction));
+                //сверка currency у 2 транзакций и работа с IntegrationService
+                transactionUpdatedEvent.Data.Value = transactionUpdatedEvent.Data.Value - prevValue.Value;
+                await _kafkaProducer.ProduceAsync("transaction-events", transaction.Id.ToString(), transactionUpdatedEvent);
+
+                return Result<TransactionResponse>.Success(transaction.ToTransactionResponse());
+            }
+
+            return Result<TransactionResponse>.Failure(TransactionErrors.TransactionNotFound);
+                
         }
 
         public async Task<bool> DeleteTransactionAsync(long id, CancellationToken token)
         {
-            await _transactionRepository.DeleteAsync(id, token);
-
-            return true;
+            var transaction = await _transactionRepository.GetByIdAsync(id, token);
+            if (transaction is not null)
+            {
+                var transactionDeletedEvent = new TransactionEvent("transaction-deleted", new TransactionData(transaction));
+                transactionDeletedEvent.Data.Value *= -1;
+                await _kafkaProducer.ProduceAsync("transaction-events", transaction.Id.ToString(), transactionDeletedEvent);
+                await _transactionRepository.DeleteAsync(id, token);
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
 
         public async Task<Result<List<TransactionResponse>>> GetTransactionsAsync(CancellationToken token)
@@ -87,6 +104,66 @@ namespace TransactionsService.Services.Implementations
                 }
                 return Result<List<TransactionResponse>>.Success(responses);
             }
+        }
+
+        public async Task<Result<List<TransactionResponse>>> GetTransactionsByUserIdAsync(long userId, CancellationToken token)
+        {
+            var transactions = await _transactionRepository.FindByCondition(l => l.UserId == userId, true).ToListAsync();
+            if (transactions is null)
+            {
+                return Result<List<TransactionResponse>>.Failure(TransactionErrors.TransactionNotFound);
+            }
+            else
+            {
+                var responses = new List<TransactionResponse>();
+                foreach (var transaction in transactions)
+                {
+                    responses.Add(transaction.ToTransactionResponse());
+                }
+                return Result<List<TransactionResponse>>.Success(responses);
+            }
+        }
+
+        public async Task<bool> DeleteTransactionsByAccountIdAsync(long accountId, CancellationToken token)
+        {
+            var transactions = await _transactionRepository.FindByCondition(l => l.AccountId == accountId, true).ToListAsync();
+            if(transactions is not null)
+            {
+                foreach (var transaction in transactions)
+                {
+                    await _transactionRepository.DeleteAsync(transaction.Id, token);
+                }
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
+
+        public async Task<bool> DeleteTransactionsByCategoryIdAsync(long categoryId, CancellationToken token)
+        {
+            var transactions = await _transactionRepository.FindByCondition(l => l.CategoryId == categoryId, true).ToListAsync();
+            if (transactions is not null)
+            {
+                foreach (var transaction in transactions)
+                {
+                    await _transactionRepository.DeleteAsync(transaction.Id, token);
+                }
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
+
+        public async Task<Result<TransactionResponse>> CreateTransactionFromAccountAsync(CreateTransactionRequest createTransactionRequest, CancellationToken token)
+        {
+            var transaction = createTransactionRequest.ToTransaction();
+            await _transactionRepository.AddAsync(transaction, token);
+
+            return Result<TransactionResponse>.Success(transaction.ToTransactionResponse());
         }
     }
 }
